+++ doc/export.rst
@@ -1,0 +1,17 @@
+
+Exporting
+================
+
+:py:mod:`mne.export`:
+
+.. automodule:: mne.export
+   :no-members:
+   :no-inherited-members:
+
+.. currentmodule:: mne.export
+
+.. autosummary::
+   :toctree: generated/
+
+   export_epochs
+   export_raw
