+++ mne/export/__init__.py
@@ -1,0 +1,1 @@
+from ._export import export_raw, export_epochs
