+++ mne/utils/docs.py
@@ -2326,7 +2326,7 @@ docdict['export_warning'] = """
 .. warning::
     Since we are exporting to external formats, there's no guarantee that all
     the info will be preserved in the external format. To save in native MNE
-    format (``.fif``) without information loss, use :func:`save` instead.
+    format (``.fif``) without information loss, use ``save`` instead.
 """
 docdict['export_params_fname'] = """
 fname : str
